feat: frontend e ajustes back

+++ backend/app/Services/AuthService.php
@@ -4,6 +4,7 @@ namespace App\Services;
 
 use App\Models\User;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Log; 
 
 class AuthService
 {
@@ -19,12 +20,17 @@ class AuthService
             $token = $user->createToken('auth_token')->plainTextToken;
 
             return response()->json([
-                'message' => 'Login realizado com sucesso',
+                'message' => 'Registro e login realizados com sucesso',
                 'access_token' => $token,
                 'token_type' => 'Bearer',
-                'user' => $user,
+                'user' => [
+                    'id' => $user->id,
+                    'name' => $user->name,
+                    'email' => $user->email,
+                ],
             ], 200);
         } catch (\Exception $e) {
+            Log::error('Erro no registro: ' . $e->getMessage(), ['trace' => $e->getTraceAsString()]);
             return response()->json([
                 'error' => 'Erro no registro: ' . $e->getMessage()
             ], 500);
@@ -43,11 +49,17 @@ class AuthService
             $token = $user->createToken('auth_token')->plainTextToken;
 
             return response()->json([
+                "message" => "Login realizado com sucesso!",
                 'access_token' => $token,
                 'token_type' => 'Bearer',
-                'user' => $user
+                'user' => [
+                    'id' => $user->id,
+                    'name' => $user->name,
+                    'email' => $user->email,
+                ]
             ], 200);
         } catch (\Exception $e) {
+            Log::error('Erro no login: ' . $e->getMessage(), ['trace' => $e->getTraceAsString()]);
             return response()->json([
                 'error' => 'Erro no login: ' . $e->getMessage()
             ], 500);
@@ -57,10 +69,14 @@ class AuthService
     public function logout($request)
     {
         try {
-            $request->user()->currentAccessToken()->delete();
-
-            return response()->json(['message' => 'Desconectado com sucesso!'], 200);
+            if ($request->user()) {
+                $request->user()->currentAccessToken()->delete();
+                return response()->json(['message' => 'Desconectado com sucesso!'], 200);
+            } else {
+                return response()->json(['message' => 'Nenhum usuário autenticado para fazer logout.'], 401);
+            }
         } catch (\Exception $e) {
+            Log::error('Erro no logout: ' . $e->getMessage(), ['trace' => $e->getTraceAsString()]);
             return response()->json([
                 'error' => 'Erro no logout: ' . $e->getMessage()
             ], 500);
